requirements.txt
# =========================================
# FLASK CORE & EXTENSIONS
# =========================================
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Admin==1.6.1
Flask-Compress==1.15

# =========================================
# AUTHENTICATION & SECURITY
# =========================================
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.8.0

# =========================================
# EMAIL & COMMUNICATION
# =========================================
Flask-Mail==0.10.0
email-validator==2.2.0

# =========================================
# UTILITIES
# =========================================
python-dotenv==1.0.1
Werkzeug==3.0.4
Pillow==11.0.0

# =========================================
# DATABASE DRIVERS
# =========================================
psycopg[binary]==3.2.3

# =========================================
# PRODUCTION SERVER
# =========================================
gunicorn==23.0.0